feat: add Task CRUD + adjust security logic and configs for all

+++ src/main/kotlin/br/pucpr/authserver/projects/ProjectService.kt
@@ -4,6 +4,9 @@ import br.pucpr.authserver.exceptions.BadRequestException
 import br.pucpr.authserver.exceptions.ForbiddenException
 import br.pucpr.authserver.exceptions.NotFoundException
 import br.pucpr.authserver.projects.requests.CreateProjectRequest
+import br.pucpr.authserver.tasks.Task
+import br.pucpr.authserver.tasks.TaskRepository
+import br.pucpr.authserver.tasks.requests.CreateTaskRequest
 import br.pucpr.authserver.users.UserRepository
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
@@ -14,18 +17,31 @@ import org.springframework.stereotype.Service
 @Service
 class ProjectService(
     val repository: ProjectRepository,
-    val userRepository: UserRepository
+    val userRepository: UserRepository,
+    val taskRepository: TaskRepository
 ) {
-    val defaultForbiddenException = ForbiddenException("User not allowed to perform this operation")
     val userNotFoundException = NotFoundException("User not found")
 
-    fun findAll() = repository.findAll()
+    fun findAll(): List<Project> = repository.findAllByOwnerId(getUserId().toLong())
 
-    fun findById(id: Long) = repository.findByIdOrNull(id) ?: throw NotFoundException("Project $id not found")
+    fun findById(id: Long): Project {
+        val project = repository.findByIdOrNull(id)
+        if (project == null || project.owner.id.toString() != getUserId()) {
+            throw NotFoundException("Project with id $id not found")
+        }
+        return project
+    }
+
+    fun findAllTasks(id: Long): List<Task> {
+        val project = repository.findByIdOrNull(id)
+        if (project == null || project.owner.id.toString() != getUserId()) {
+            throw NotFoundException("Project with id $id not found")
+        }
+        return project.tasks
+    }
 
     fun insert(projectRequest: CreateProjectRequest): Project {
-        val userId = SecurityContextHolder.getContext().authentication?.name
-            ?: throw defaultForbiddenException
+        val userId = getUserId()
         val owner = userRepository.findById(userId.toLong()).orElseThrow { userNotFoundException }
         val project = projectRequest.toProject(owner)
 
@@ -36,13 +52,21 @@ class ProjectService(
             .also { log.info("Project with id ${project.id} created successfully") }
     }
 
-    fun update(id: Long, projectRequest: CreateProjectRequest): Project {
-        val userId = SecurityContextHolder.getContext().authentication?.name
-            ?: throw defaultForbiddenException
-        val project = repository.findById(id)
-            .orElseThrow { NotFoundException("Project with id $id not found") }
+    fun insertTask(id: Long, taskRequest: CreateTaskRequest): Task {
+        val project = repository.findByIdOrNull(id)
+        if (project == null || project.owner.id.toString() != getUserId()) {
+            throw NotFoundException("Project with id $id not found")
+        }
+        val task = taskRequest.toTask(project)
+        return taskRepository.save(task)
+            .also { log.info("Task ${task.title} created successfully for project $id") }
+    }
 
-        if (project.owner.id.toString() != userId) throw defaultForbiddenException
+    fun update(id: Long, projectRequest: CreateProjectRequest): Project {
+        val project = repository.findByIdOrNull(id)
+        if (project == null || project.owner.id.toString() != getUserId()) {
+            throw NotFoundException("Project with id $id not found")
+        }
         project.name = projectRequest.name!!
         project.description = projectRequest.description!!
         return repository.save(project)
@@ -50,19 +74,21 @@ class ProjectService(
     }
 
     fun delete(id: Long) {
-        val userId = SecurityContextHolder.getContext().authentication?.name
-            ?: throw defaultForbiddenException
+        val userId = getUserId()
         val user = userRepository.findById(userId.toLong()).orElseThrow { userNotFoundException }
-        val project = repository.findById(id)
-            .orElseThrow { NotFoundException("Project with id $id not found") }
-        if (!user.isAdmin() && project.owner.id.toString() != userId) {
-            throw defaultForbiddenException
+        val project = repository.findByIdOrNull(id)
+        if (project == null || !user.isAdmin() && project.owner.id.toString() != userId) {
+            throw NotFoundException("Project with id $id not found")
         }
         repository.deleteById(id)
             .also { log.warn("Project with id {} deleted successfully", id) }
     }
 
     companion object {
+        val defaultForbiddenException = ForbiddenException("User not allowed to perform this operation")
         val log: Logger = LoggerFactory.getLogger(ProjectService::class.java)
+
+        fun getUserId() = SecurityContextHolder.getContext().authentication?.name
+            ?: throw defaultForbiddenException
     }
 }
